\name{BSCRN-package}
\alias{BSCRN-package}
\alias{BSCRN}
\docType{package}
\title{
\packageTitle{BSCRN}
}
\description{
\packageDescription{BSCRN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BSCRN}
\packageIndices{BSCRN}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BSCRN}

Maintainer: \packageMaintainer{BSCRN}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
