% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{CVBFtestrsplit}
\alias{CVBFtestrsplit}
\title{Title}
\usage{
CVBFtestrsplit(dataset1, dataset2, trainsize1, trainsize2, seed = NULL)
}
\arguments{
\item{dataset1}{One dataset that we want to check if it has the same distribution as another data set}

\item{dataset2}{Another dataset that we want to check if it has the same distribution as another data set}

\item{trainsize1}{The training set size of dataset 1}

\item{trainsize2}{The training set size of dataset 2}

\item{seed}{The seed used to generate training set and validation sets for both of the data sets}
}
\value{
A log BF that tests whether two distributions are the same via CVBF.
}
\description{
Title
}
