% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisettetrainlabs}
\alias{gisettetrainlabs}
\title{Labels for data that represent 4 or 9 in image recognition. It is unknown which label corresponds to which image. This data set was given as a blind classification problem by NIPS in 2003. The competition involves guessing what some validation labels were. These were the training labels ued for statisticains to train their models before submission.}
\format{A named matrix with 6000 rows and 1 column, each value is -1 or 1 and corresponds to whether or not an image is a 4 or 9 in the training data set.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Gisette}
}
\usage{
gisettetrainlabs
}
\description{
Labels for data that represent 4 or 9 in image recognition. It is unknown which label corresponds to which image. This data set was given as a blind classification problem by NIPS in 2003. The competition involves guessing what some validation labels were. These were the training labels ued for statisticains to train their models before submission.
}
\keyword{datasets}
