% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisettetrainpreds}
\alias{gisettetrainpreds}
\title{Predictors of the labels in gisettetrainlabs. Some of these correspond to pixel intensity measurements of some images, while others are random noise. It is uncertain which are noisy probes and which are pixel measurements.}
\format{A named matrix with 6000 rows and 5000 columns. It is zero inflated, as not every pixel is touched when a person writes a number.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Gisette}
}
\usage{
gisettetrainpreds
}
\description{
Predictors of the labels in gisettetrainlabs. Some of these correspond to pixel intensity measurements of some images, while others are random noise. It is uncertain which are noisy probes and which are pixel measurements.
}
\keyword{datasets}
