% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisettevalidlabs}
\alias{gisettevalidlabs}
\title{More labels for data that represent 4 or 9 in image recognition. It's unknown which label corresponds to which image. This data set was given as a blind classification problem by NIPS in 2003. The competition involves guessing what the validation labels were. These were released after the competition.}
\format{A named matrix with 1000 rows and 1 column.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Gisette}
}
\usage{
gisettevalidlabs
}
\description{
More labels for data that represent 4 or 9 in image recognition. It's unknown which label corresponds to which image. This data set was given as a blind classification problem by NIPS in 2003. The competition involves guessing what the validation labels were. These were released after the competition.
}
\keyword{datasets}
