% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gisettevalidpreds}
\alias{gisettevalidpreds}
\title{Predictors of the labels in gisettevalidlabs. Some of these correspond to pixel intensity measurements of some images, while others are random noise. It is uncertain which are noisy probes and which are pixel measurements.}
\format{A named matrix with 1000 rows and 5000 columns.}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Gisette}
}
\usage{
gisettevalidpreds
}
\description{
Predictors of the labels in gisettevalidlabs. Some of these correspond to pixel intensity measurements of some images, while others are random noise. It is uncertain which are noisy probes and which are pixel measurements.
}
\keyword{datasets}
