% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{HallKernel}
\alias{HallKernel}
\title{Compute Hall Kernel density estimate, given a training set and a set of values to evaluate on}
\usage{
HallKernel(h, datagen2, x)
}
\arguments{
\item{h}{Bandwidth parameter}

\item{datagen2}{Training set for Hall KDE}

\item{x}{Value (or values) to evaluate KDE on. Can be a scalar, a matrix, or a vector.}
}
\value{
An object of the same dimension as x that evaluates a KDE trained on datagen2 on everyone of the points in x
}
\description{
Compute Hall Kernel density estimate, given a training set and a set of values to evaluate on
}
