% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{integrand.Hall}
\alias{integrand.Hall}
\title{Compute integrand of the Marginal likelihood for the CVBF}
\usage{
integrand.Hall(h, y, x, cons, hhat)
}
\arguments{
\item{h}{Bandwidth parameter, this is what the integral is with respect to.}

\item{y}{Validation set}

\item{x}{Training set}

\item{cons}{The integrand is typically too big to easily evaluate, need a constant to help evaluate it. Typically just use the log version of this function instead}

\item{hhat}{The parameter to center the prior at, recommended to be maximizer of the likelihood}
}
\value{

}
\description{
Compute integrand of the Marginal likelihood for the CVBF
}
