% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{KHall}
\alias{KHall}
\title{Title Hall Kernel}
\usage{
KHall(x)
}
\arguments{
\item{x}{The parameter to evaluate the Hall Kernel density on}
}
\value{
Evaluation of K(X), where $K \propto exp^{-.5log(1 + abs(x))^2}$ and is a valid PDF
}
\description{
Title Hall Kernel
}
