% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{logintegrand.Hall}
\alias{logintegrand.Hall}
\title{Compute log of the integrand of the Marginal likelihood for CVBF}
\usage{
logintegrand.Hall(h, y, x, hhat)
}
\arguments{
\item{h}{Bandwidth parameter, this is the variable that is being integrated over}

\item{y}{Validation set to evaluate the likelihood over}

\item{x}{Training set to build the KDE}

\item{hhat}{Parameter that specifies where prior should be centered}
}
\value{
Evaluation of the integrand at a particular log likelihood value.
}
\description{
Compute log of the integrand of the Marginal likelihood for CVBF
}
