% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{loglike.KHall}
\alias{loglike.KHall}
\title{Evaluate a log likelihood using the Hall Kernel}
\usage{
loglike.KHall(h, y, x)
}
\arguments{
\item{h}{A bandwidth or vector of bandwidths tot ry out}

\item{y}{A validation set to evaluate the KDE on}

\item{x}{A training set to build the Hall kernel density estimate on}
}
\value{
-log likelihood evaluation, where the likelihood is constructed using the training data
}
\description{
Evaluate a log likelihood using the Hall Kernel
}
\examples{
dataset1 = rnorm(100)
DT = dataset1[1:50]
DV = dataset1[51:100]
loglike.Khall(.01, DT, DV)
}
