% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{logpriorused}
\alias{logpriorused}
\title{Evaluate the log prior on one of the data sets}
\usage{
logpriorused(h, hhat)
}
\arguments{
\item{h}{A value to evaluate the prior on}

\item{hhat}{Tuning parameter for the prior}
}
\value{
Evaluates a particular type of prior placed on the bandwidth for CVBF.
}
\description{
Evaluate the log prior on one of the data sets
}
