% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTreeFunctions.R
\name{PolyaTreeBFcons}
\alias{PolyaTreeBFcons}
\title{Polya Tree test using constructed objects}
\usage{
PolyaTreeBFcons(PolyaTreePriorLik1, PolyaTreePriorLik2)
}
\arguments{
\item{PolyaTreePriorLik1}{An object constructed by PolyaTreePriorLikCons for a dataset.}

\item{PolyaTreePriorLik2}{Another object constructed by PolyaTreePriorLikCons for another dataset.}
}
\value{
A scalar, which corresponds to the log BF of the test. Another vector is given which corresponds to the contribution of the log BF at each level.
}
\description{
Polya Tree test using constructed objects
}
\examples{
set.seed(100)
dataset1 = rnorm(200)
dataset2 = rnorm(200)
PTmodel1 = PolyaTreePriorLikCons(datasetX = dataset1, Ginv = qnorm)
PTmodel2 = PolyaTreePriorLikCons(datasetX = dataset2, Ginv = qnorm)
PTresults = PolyaTreeBFcons(PTmodel1, PTmodel2)
}
