% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTreeFunctions.R
\name{PolyaTreePredDraws}
\alias{PolyaTreePredDraws}
\title{Generate draws from a Polya Tree model's Predictive distribution}
\usage{
PolyaTreePredDraws(PolyaTreePriorLik, ndraw = 2000)
}
\arguments{
\item{PolyaTreePriorLik}{An object constructed by PolyaTreePriorLikCons for a dataset.}

\item{ndraw}{Number of draws desired from Predictive Polya Tree distribution.}
}
\value{
A list of draws from the Predictive Polya Tree Posterior
}
\description{
Generate draws from a Polya Tree model's Predictive distribution
}
\examples{
set.seed(100)
dataset1 = rnorm(200)
PTmodel1 = PolyaTreePriorLikCons(datasetX = dataset1, Ginv = qnorm)
Predictiveposteriordraws = PolyaTreePredDraws(PTmodel1, ndraw = 400)
plot(density(Predictiveposteriordraws))
}
