% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTreeFunctions.R
\name{PolyaTreePriorLikCons}
\alias{PolyaTreePriorLikCons}
\title{Construct a Polya tree object for a data set}
\usage{
PolyaTreePriorLikCons(datasetX, Ginv = NULL, c = 1, leveltot = NULL)
}
\arguments{
\item{datasetX}{A dataset to compute the Polya Tree prior on}

\item{Ginv}{A quantile function of some distribution, use to make bins}

\item{c}{A scalar. The higher this is, the more influential the prior is on the data set.}

\item{leveltot}{The number of levels the Polya Tree should go down.}
}
\value{
A list of vectors called alphalist, a list of vectors called splitlist, c, leveltot, and Ginv. Alphalist is used to construct a prior. Splitlist is used to construct a likelihood. Call the collection of these a Polya Tree object.
}
\description{
Construct a Polya tree object for a data set
}
\examples{
set.seed(100)
dataset1 = rnorm(200)
PTmodel1 = PolyaTreePriorLikCons(datasetX = dataset1)
#Ptmodel1 can be called by other methods in the package to use it
}
