% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{PredCVBFIndepMHbw}
\alias{PredCVBFIndepMHbw}
\title{Draw bandwidths from CVBF predictive posterior by independent Metropolis Hasting sampling}
\usage{
PredCVBFIndepMHbw(
  ndraw = 100,
  propsd = NULL,
  maxIter = 10000,
  XT1,
  XV1,
  startingbw = NULL
)
}
\arguments{
\item{ndraw}{Number of unique draws desired for the bandwidth parameter from the posterior.}

\item{propsd}{A tuning parameter, corresponds to what proposal standard deviation should be for when using MH to traverse the posterior. Should be chosen with care to ensure good mixing. Higher acceptance rates are OK here, compared to classic MH.}

\item{maxIter}{The max number of MH iterations to try. Do not set to be too large. It will kick the code out if acceptance rates for MH are small.}

\item{XT1}{Training set for a data set}

\item{XV1}{Validation set for a data set}

\item{startingbw}{A value to start the MH chain at. If not provided, starts at posterior mode. All proposals will be drawn from a distribution whose center is startingbw. This is normally a bad idea, but the posterior is some type of unimodal distribution, so this is actually effective.}
}
\value{
A list of bandwidths that come from the posterior distribution. This will be larger than ndraw, as some draws will be repeats.
}
\description{
Draw bandwidths from CVBF predictive posterior by independent Metropolis Hasting sampling
}
