% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreeningMethods.R
\name{SeqScreenVars}
\alias{SeqScreenVars}
\title{Screen a data set for important functions in parallel}
\usage{
SeqScreenVars(
  datasetX,
  datasetY,
  method = "SIS",
  cutoff = NULL,
  train1ids = NULL,
  trainsize1 = NULL,
  trainsize2 = NULL,
  train2ids = NULL,
  seed = NULL,
  Ginv = NULL,
  c = NULL,
  leveltot = NULL,
  PTscale = TRUE
)
}
\arguments{
\item{datasetX}{A matrix containing values that are predictors for the Y values}

\item{datasetY}{A vector containg the class that each predictor corresponds to. For now can only handle binary responses.}

\item{method}{A string containing the type of screening to do. Can be "SIS", "KS", "CVBF" or "PT"}

\item{cutoff}{A real number that corresponds either to an alpha value for testing or a cutoff value on how large the Bayes factor needs to be to conclude a difference exists.}

\item{train1ids}{A vector of ids that correspond to which observations to use for the training set for the first data set}

\item{trainsize1}{Size of the training set for one of the classes for CVBF}

\item{trainsize2}{Size of the training set for the other one of the classes for CVBF}

\item{train2ids}{A vector of ids that correspond to which observations to use for the training set for the second data set}

\item{seed}{A seed for CVBF based screening, can use this to reproduce results instead of train1ids or train2ids.}

\item{Ginv}{A function to compute quantiles with for Polya tree.}

\item{c}{Tuning parameter for Polya tree, signifies how impactful prior should be.}

\item{leveltot}{Depth of Polya tree to construct if Polya tree based screening is type of screening chosen}

\item{PTscale}{A True / false variable. Should columns be standardized before proceeding with Polya tree based screening? Default is to screen as recommended by authors.}
}
\value{
A list of variables that are interpreted to be important
}
\description{
Screen a data set for important functions in parallel
}
\examples{
data(gisettetrainlabs)
data(gisettetrainpreds)
ImpVarsSIS1 = SeqScreenVars(datasetX = gisettetrainpreds[, 1:500], datasetY = gisettetrainlabs[,1], method = "SIS")
length(ImpVarsSIS1$varspicked)
ImpVarsKS1 = SeqScreenVars(datasetX = gisettetrainpreds[, 1:500], datasetY = gisettetrainlabs[,1], method = "KS")
length(ImpVarsKS1$varspicked)
ImpVarsPT1 = SeqScreenVars(datasetX = gisettetrainpreds[, 1:500], datasetY = gisettetrainlabs[,1], method = "PT", c = 1, leveltot = 12, Ginv = qnorm, PTscale = TRUE)
#Only do on first 500
length(ImpVarsPT1$varspicked)
hist(ImpVarsPT1$logBFlist)
ImpVarsCVBF1 = SeqScreenVars(datasetX = gisettetrainpreds[, 1:500], datasetY = gisettetrainlabs[,1], method = "CVBF", trainsize1 = 2960, trainsize2 = 2960, seed = 200)
length(ImpVarsCVBF1$varspicked)

}
