\name{BSCRN-package}
\alias{BSCRN-package}
\alias{BSCRN}
\docType{package}
\title{
\packageTitle{BSCRN}
}
\description{
\packageDescription{BSCRN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BSCRN}
\packageIndices{BSCRN}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BSCRN}

Maintainer: \packageMaintainer{BSCRN}
}
\references{
\href{https://orfe.princeton.edu/~jqfan/papers/06/SIS.pdf}{Original paper on SIS}: Fan, J., & Lv, J. (2008). Sure independence screening for ultrahigh dimensional feature space. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 70(5), 849-911.

\href{https://watermark.silverchair.com/ass062.pdf?token=AQECAHi208BE49Ooan9kkhW_Ercy7Dm3ZL_9Cf3qfKAc485ysgAAAnkwggJ1BgkqhkiG9w0BBwagggJmMIICYgIBADCCAlsGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMH1cTkqIl5XaRu4GOAgEQgIICLEycgs8ydDE8p1JNs74TA8gk52VVacMnU0_TGLeLyiHHV4esncCSiBbIZ1WoeU3I4WLgn836LIest9MaFfpwAU9ywSoYNx5uwMjOHCrouOyNI10cSzcK9HCXH9iP2r5J6uhDQQX7Qo8sa0HUs-Wa-BgDRqOk1geA94xZuID0pZCUua6ehZ_WW08OhLX8yE0jBMNkCZb73k7nNMxJQMsNvjJig3M_XhxvTuXh3j4MYuIUJfQjKibcTNFjBlEfh4g_XsDTlKS3VnxTwO-nW4VrmYiXELf3VHxm1C4wrSccJgsh0Hvs-bms3e0I2onbfcVj14uOUnMnAJRQHv4H4tQed2N9cWF9SaW0oXcPniuo873PQh6OVlBM3pUSQ9EKE6U7Gu_WRRo2Wb6ZnOAvvSbmt-oolHQeMeUTucGsvPwEIKKP9Zkrk-ietvSnQUsOnD195UzCU3lByeED7RcOawMA-ytc6VHj-qRpotKj8pXF8bPcKFtoyk-5h1Y9Bz6ZaJ533GfPxO145RONQZANbSjP2a--PmyLylmKaIkfKQy0ZO_iUBydRACkzfhQKFYSJzT6-7UdMUdBx3Ln0GEqhjar05XWeZMr1AkBdCf5GNwga-Hsrk5OaqNpGqW3VIgs-IVOeRzlfqTsMYyG1TG3dSVdexvsrVKs1OuDWUxtWmKUalvkHvCT3vIR1HGPatnLig5Hl1thQ_JyaB-nGsWnQ5NMEtRggjC-wAGT7gL3dK8}{First paper on Kolmogorov Smirnov Screening}: Mai, Q., & Zou, H. (2012). The Kolmogorov filter for variable screening in high-dimensional binary classification. Biometrika, 100(1), 229-234.
~~ Literature or other references for background information ~~

\href{https://www.jstor.org/stable/27639772?seq=1#metadata_info_tab_contents}{Hanson's paper on inference of Polya Trees} Hanson, T. (2006). Inference for Mixtures of Finite Polya Tree Models. Journal of the American Statistical Association, 101(476), 1548-1565. Retrieved from www.jstor.org/stable/27639772

\href{Holmes et al.'s POlya Tree test}{https://arxiv.org/pdf/0910.5060.pdf} Holmes, C. C., Caron, F., Griffin, J. E., & Stephens, D. A. (2015). Two-sample Bayesian nonparametric hypothesis testing. Bayesian Analysis, 10(2), 297-320.

\href{https://reader.elsevier.com/reader/sd/pii/S0167947312003945?token=1050252D85E674DBD37024EBCF00550AB0972473F7CDE3A6BAB5AF4CDC141C20C6FC47E94CB61E36C89A4E79575CD634}{Hanson and Chen's Polya Tree test} Chen, Y., & Hanson, T. E. (2014). Bayesian nonparametric k-sample tests for censored and uncensored data. Computational Statistics & Data Analysis, 71, 335-346.

CVBF manuscript is in preperation, please contact author for current manuscript.
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
