% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{CVBFtestrsplit}
\alias{CVBFtestrsplit}
\title{Compute a CVBF, a Bayes factor that checks if two data sets share the same distribution.}
\usage{
CVBFtestrsplit(
  dataset1,
  dataset2,
  trainsize1,
  trainsize2,
  seed = NULL,
  train1_ids = NULL,
  train2_ids = NULL
)
}
\arguments{
\item{dataset1}{One dataset that we want to check if it has the same distribution as another data set}

\item{dataset2}{Another dataset that we want to check if it has the same distribution as another data set}

\item{trainsize1}{The training set size of dataset 1}

\item{trainsize2}{The training set size of dataset 2}

\item{seed}{The seed used to generate training set and validation sets for both of the data sets}

\item{train1_ids}{Indices for the training set of dataset 1}

\item{train2_ids}{Indices for the training set of dataset 2}
}
\value{
A list containing a log BF that tests whether two distributions are the same via CVBF, and the training sets used to generate the CVBF.
}
\description{
Compute a CVBF, a Bayes factor that checks if two data sets share the same distribution.
}
\examples{
set.seed(100)
dataset1 = rnorm(200)
dataset2 = rnorm(200) 
CVBF1 = CVBFtestrsplit(dataset1, dataset2, trainsize1 = 100, trainsize2 = 100)
CVBF1$logBF #Gives back the log Bayes factor
CVBF1$train1_ids #Gives back the training set of the first data set
CVBF1$train2_ids #Gives back the training set of the second data set

}
