% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{KHall}
\alias{KHall}
\title{Title Hall Kernel evaluation}
\usage{
KHall(x)
}
\arguments{
\item{x}{The parameter to evaluate the Hall Kernel density on}
}
\value{
Evaluation of K(X), where K is a heavy tailed kernel. For more information on K(x), see
 K_0 on Hall's paper on Kullback Leibler loss (Annals of Statistics 1957)
}
\description{
Title Hall Kernel evaluation
}
\examples{
Khall(.1)
}
