% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{laplace.kernH2c}
\alias{laplace.kernH2c}
\title{Compute Marginal Likelihoods for CVBF}
\usage{
laplace.kernH2c(y, x, hhat, c)
}
\arguments{
\item{y}{Validation Set}

\item{x}{Training set}

\item{hhat}{Bandwidth parameter that maximizes the log likelihood}

\item{c}{A constant that is equal to the log likelihood + log prior evaluated at the maximum}
}
\value{
Evaluation of the CVBF marginal likelihood via Laplace Approximation.
}
\description{
Compute Marginal Likelihoods for CVBF
}
