% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTreeFunctions.R
\name{PolyaTreePredDraws}
\alias{PolyaTreePredDraws}
\title{Title}
\usage{
PolyaTreePredDraws(PolyaTreePriorLik, ndraw = 2000)
}
\arguments{
\item{PolyaTreePriorLik}{An object constructed by PolyaTreePriorLikCons for a dataset.}

\item{ndraw}{Number of draws desired from Predictive Polya Tree distribution.}
}
\value{
A list of draws from the Predictive Polya Tree Posterior
}
\description{
Title
}
