% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVBFFunctions.R
\name{PredCVBFDens}
\alias{PredCVBFDens}
\title{Title}
\usage{
PredCVBFDens(bwvec, XT1)
}
\arguments{
\item{bwvec}{A vector of bandwidths, can come from either of the methods that draw bandwidths from the posterior.}

\item{XT1}{The training set}
}
\value{
A function that evaluates the predictive posterior at particular values
}
\description{
Title
}
